added intersection observer
I added intersection observers to watch each of the major sections. The ammount of the section seen on the screen controls how green the navbar background is for that section. The ids of teh navbar and of the sections need to match. I also changed the navbar to a class instead of an id. I also changed the hover style of the narbar links for over and under line.

I cannot think of any other feature changes that I need to make. The next step is to update the content. Then I am done.