updated js and favicons
created a 16x16 adn 32x32 icon for the favicon (small image on tab bar).

also updated teh js. it now handles the hover and clicking on the menu when the screen is small and the menu consolidates into a dropdown. there is a bit of trickery with a flag and event queue timing in order for it to work.