updated colors, altered layout, started to fill content
I started to copy over content from the old sites. Not everything was added, and no spell check was used. This needs to be revisited.

The colors were updated and adjusted. named colors are now used instead of numbers.

A few new layouts were added, the problem right now is that the images are too big and there is not enouth content in the textboxes that go with the images. The width ratios could be further adjusted.