all info has been added
all info has been added and spell checked. the only thing left to do is update the resume.

updated the js to better handle highlighting on the navbar. the intersection observer was not functioning exactly like I hoped for. with some math i was able to fix it.

i updated the css with a different padding for each 'testrow'

i also uploaded a few work docs to the docs folder. one is my resume and the other is the info used in the website. the resume should stay here and be updated periodically. the about me should be deleted after the resume is updated.