finialized updates
updated a few last things and moved the files around. The old site that used to be there is now in the folder oldersite. the site that I was working on before this new one (but i never finished) is is oldsite (used to be named newsite). contents from newnewsite (which i have been working on for the past few weeks) have been copied into the base folder.